Cache: Extract interaction behavior to Adapter class
