+++ lib/site/adapter.rb
@@ -1,0 +1,11 @@
+module Site
+
+	class Adapter
+
+		def initialize(env)
+			@env ||= env
+		end
+
+	end
+
+end
