+++ lib/site/cache.rb
@@ -1,7 +1,6 @@
 require 'fileutils'
 require 'digest'
 
-require 'redis'
 require 'listen'
 require 'mime-types'
 require 'tilt'
@@ -9,6 +8,9 @@ require 'tilt'
 require 'site/cache/event'
 require 'site/cache/entry'
 require 'site/logger'
+
+require 'site/adapter'
+require 'site/redis_adapter'
 
 module Site
 
@@ -19,27 +21,7 @@ module Site
 
 			@env, @application = env, application
 
-			redis_host = ENV["REDIS_HOST"]
-			redis_port = ENV["REDIS_PORT"]
-			redis_password = ENV["REDIS_PASSWORD"]
-
-			redis_opts = {}
-
-			redis_opts[:host] = redis_host if redis_host
-			redis_opts[:port] = redis_port if redis_port
-			redis_opts[:password] = redis_password if redis_password
-
-			redis_opts.tap do |hash|
-				filter_keys = [:password]
-
-				printable_opts = hash.map do |key, value|
-					filter_keys.include?(key) ? [key, "[FILTERED]"] : [key, value]
-				end.to_h
-
-				Logger.info "cache" do
-					"Connecting to redis with opts #{printable_opts}"
-				end
-			end
+			@adapter = RedisAdapter.new @env
 
 			@listener = Listen.to(Site::STATIC_DIRECTORY, Site::VIEWS_DIRECTORY) do |modified, added, removed|
 				dispatch(modified, added, removed)
@@ -51,8 +33,6 @@ module Site
 
 			warm(Site::STATIC_DIRECTORY)
 			warm(Site::VIEWS_DIRECTORY)
-
-			@redis = Redis.new redis_opts
 		end
 
 		def dispatch(modified, added, removed)
@@ -72,47 +52,22 @@ module Site
 		end
 
 		def handle_event(event)
-
 			case event
 			when RemovedEvent
 				handle_delete event
 			when AddedEvent, ModifiedEvent
 				handle_modified event
 			end
+		end
 
+		def get(key)
+			@adapter.get(key)
 		end
 
 		def set_entry(entry)
 			slug = { sha: entry.encoded_contents, data: entry.contents }
 
-			store(entry.filename, slug)
-		end
-
-		def store(key, object)
-			data = JSON.generate(object)
-			@redis.set key, data
-		end
-
-		def contains?(key)
-			@redis.exists key
-		end
-
-		def get(key)
-			data = @redis.get key
-			JSON.parse(data) if data
-		end
-
-		def dump!
-			Logger.info "cache#dump!" do "Beginning dump!" end
-
-			keys = @redis.keys
-
-			Logger.info "cache#dump!" do "Have #{keys.count} keys..." end
-
-			keys.each do |key|
-				value = @redis.get key
-				Logger.debug "cache#dump!" do "  #{key} => #{value}" end
-			end
+			@adapter.store(entry.filename, slug)
 		end
 
 		protected
@@ -120,10 +75,10 @@ module Site
 		def handle_delete(event)
 			entry = Entry.new(event.filename)
 
-			if contains?(entry.filename)
+			if @adapter.contains?(entry.filename)
 				Logger.debug "cache#handle_event" do "#{entry.relative_path_from_root}: removing from cache, route delete" end
 				t0 = Time.now
-				@redis.delete entry.filename
+				@adapter.delete entry.filename
 				@application.routes_delete(entry.routes)
 				t1 = Time.now
 				Logger.debug "cache#handle_event" do "ok (#{t1 - t0}s)" end
@@ -135,8 +90,8 @@ module Site
 		def handle_modified(event)
 			entry = Entry.new(event.filename)
 
-			if contains?(entry.filename)
-				slug = get(entry.filename)
+			if @adapter.contains?(entry.filename)
+				slug = @adapter.get(entry.filename)
 
 				if slug["sha"] == entry.encoded_contents
 					Logger.debug "cache#handle_event" do "#{entry.relative_path_from_root}: already in cache; no change" end
